multipage pdf tranformer to mds

+++ lib/database/local_mind_base.dart
@@ -2,6 +2,7 @@
 
 import 'dart:io';
 
+import 'package:flutter/material.dart';
 import 'package:ini/ini.dart';
 import 'package:mind_base_manager/database/mind_base.dart';
 import 'package:mind_base_manager/database/mind_base_md_converter.dart';
@@ -50,6 +51,9 @@ class LocalMindBase extends MindBase {
     final List<FileSystemEntity> entities = await dir.list().toList();
     final Iterable<File> files = entities.whereType<File>();
     for (var v in files) {
+      if (!v.path.contains(".md")) {
+        continue;
+      }
       if (v.path.contains("/.md")) {
         v.delete();
         continue;
@@ -209,6 +213,15 @@ class LocalMindBase extends MindBase {
       config.set("config", "id", studentMetadata.id);
       File("userdata/config.ini").writeAsString(config.toString());
     });
+  }
+
+  @override
+  Image image(String id) {
+    return Image.file(
+        File(
+          "$pathMdFiles/$id",
+        ),
+        fit: BoxFit.fitWidth);
   }
 }
